finish transactions

+++ app/core/model/transaction_dto.py
@@ -1,10 +1,11 @@
+from decimal import Decimal
 from dataclasses import dataclass
 
 
 @dataclass
 class TransactionDTO:
-    amount: float
-    commission: float
+    amount: Decimal
+    commission: Decimal
     coin_type_id: int
     wallet_from_address: str
     wallet_to_address: str
